makes notebook output more distinct from docstring
adds a notebook arg to iter_extract_docstring_from_text, which
determines whether to find the lineno or just use 'n/a', since it
doesn't really make sense to get the lineno from markdown cells

this is used to determine whether the short_purpose should say it came
from parsing a docstring or a markdown cell.

making all the linenos n/a meant there was an issue with extracting
multiple dois from the same file. to fix that, added "{name}" to the
reference identifier